+++ internal/tui/delete_confirm_popup_model.go
@@ -76,7 +76,7 @@ func (m DeleteConfirmPopupModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.KeyMsg:
 		switch {
 		case key.Matches(msg, m.keys.Quit, m.keys.GlobalQuit):
-			return m, tea.Quit
+			return m, programQuitCmd()
 		case key.Matches(msg, m.keys.Esc):
 			return m, func() tea.Msg { return closePopupMsg{} }
 		case key.Matches(msg, m.keys.Enter):